Improve Result's Display impl for extension values
Previously, the `Display` impl for `vk::Result` did not include handling
for extension values. For example, VK_ERROR_OUT_OF_DATE_KHR would
display simply as "-1000001004". Now, we fall back to the `Debug` impl
of the `Result` if the value is unknown (i.e. from an extension). This
preserves the current nice messages for non-extension values, and the
numeric output for truly unknown values, but displays the enum variant
name (e.g. "ERROR_OUT_OF_DATE")  for extension values.