Give users direct access to bitflags and enumeration constants
Nothing special happens when converting to and from these constants, and
its API is found to be more concise yet ever so slightly (0.1s) faster
to compile than having separate `to_raw` and `from_raw` functions for
each, that it makes sense to drop these entirely.  Ash is still on a
nonbreaking contribution streak so these functions are marked
`deprecated` for now and will be removed at the next breaking crate
release.

[1]: https://github.com/MaikKlein/ash/pull/549#discussion_r781402363