generator: Skip setters on `returnedonly="true"` structs
Vulkan annotates structs that are purely returned by the driver and
do not have to be constructed by users (besides initializing `sType`/
`pNext`).  For these we can skip all builder functions (except
`push_next()` and the `CStr`/slice getter helpers) and lighten our
codebase somewhat.

Note that it is still possible to update the structure in a more
low-level way by directly acessing the fields (or via FRU syntax).
For Rust-based layer implementations this may be somewhat cumbersome,
for which we could re-expose the setters behind a `cfg(feature =
"returnedonly-setters")` of sorts?