Work around `unused_qualifications` lint for Rust 1.80 prelude extension
`size_of(_val)()` was added to the prelude in Rust 1.80, causing
`unused_qualifications` warnings whenever we import `mem` in scope and
call it with `mem::` prefix.

The easiest workaround is to remove the prefix and explicitly import
the function(s) in scope, which was already done in a few places.  We
annotate the import with a `TODO` to remove it once bumping our MSRV
on or past 1.80.