generator: Print `vk-parse` errors when they're encountered (#930)
Some upstream `vk.xml` changes are causing failures on missing registry
children later on in our generator code.  These are complicated to
debug, unless we print the errors that `vk-parse` already reports.  In
this debug case it was clarifying that some new and unknown elements
upstream are the reason for high-level registry elements to not be
parsable.