generator: Demote `vk-parse` error-asserts to a warning message
When `vk-parse` encounters XML parsing errors (e.g. on unknown attributes
or elements) it skips the element and emits an error in a list of errors
that are returned to the caller.  We were originally never checking
these leading to hard-to-debug issues, but eventually started asserting
on this list being empty or otherwise panicking in #930.

Since `ash`'s generator is used in Khronos' upstream CI (in a fallible
yet warning way), any new attribute would cause the generator to fail
and subsequently require updates to `vk-parse` and `ash` before their
CI is no longer flagged as "succeeded with warnings".  "Solve" this by
once again allowing errors to exist, while still at least printing them
to `stderr` for insights (that were previously missing...) to anyone
running the generator.

Note that this once again allows the generator to fail further in the
chain when expected elements are missing, in which case an update is
required either way.  If this happens often we can adjust `vk-parse` to
be more lenient in still emitting errors but continuing to parse said
element, if the new/unrecognized elements or attributes are considered
harmless to the parsing process.