generator: Elide lifetimes in `impl` blocks whenever they are redundant
Rust 1.83 suggests that whenever a named lifetime `'a` in an `impl<'a>
(Trait for) Struct<'a>` is specified, but not used anywhere else within
that `impl` block, the naming can be removed and replaced with `'_`.

While we can unconditionally remove this from blocks like `impl
TaggedStructure`, more generator logic is necessary to optionally omit
them on builder blocks with special care around optional presence of
`push_next()`.