generator: Omit lifetime on `vk::AllocationCallbacks` for opaque pointer
When a struct only contains pointers to opaque types (only
`vk::AllocationCallbacks` was found to match this predicate), which
we don't track in the builder, the lifetime parameter is completely
unnecessary and useless as we don't have anything useful to key this off
of.  The builder shouldn't force the incoming pointer to be any specific
Rust thing to track the lifetime through, especially since its lifetime
protrudes the lifetime of the `vk::AllocationCallbacks` structure
and is passed into the callbacks whenever Vulkan wishes to call them,
which could happen after an API command completes and a (stack-local)
`vk::AllocationCallbacks` instance is destroyed again.