Change `push_next()` to `unsafe fn extend()` and add `push()` alternative (#909)
`push_next()` is totally `unsafe` because of dereferencing a chain of
`p_next` pointers to find the end of the chain to insert, which was
obfuscated by a large `unsafe` block for the `BaseOutStructure` pointer
cast in commit c8c8f69a ("`next` can contain a pointer chain and we need
to correct insert it.").

While this function should definitely be marked unsafe, wrapping
builders in `unsafe {}` en masse in user code isn't all too desirable,
especially when this soundness issue only exists to optionally walk
a `p_next` chain while most users are likely inserting bare structs
without pointer chains most of the time.  `push()` is introduced for
this reason, remaining safe to call without any unintended raw pointer
dereferences.  `push_next()` has been renamed to `extend()` to more
closely align with standard naming, and force existing users to notice
this change and hopefully analyze their uses to decide which are
safe/correct to switch over to safe `push()`.