Drop unnecessary trailing semicolons in function implementations
This is the inverse of the `clippy::semicolon_if_nothing_returned` lint,
where intentionally-omitted semicolons ensure at compile-time that the
return type of the C function is identical to what we return out of the
high-level wrapper, typically a `()` void type.  This makes it harder to
accidentally forget to return a count value for example.  `vk::Result`
is already annotated with `#[must_use]` however, and needs a specific
conversion to be turned into the `VkResult = Result` alias.