examples: Implement triple-buffering
This removes a per-frame stall and demonstrates how users should
utilize Vulkan to keep the GPU fed with rendering work, as long as that
work (and corresponding presentation requests) complete before we're
getting 3 frames ahead.

Note that this does not implement proper frame throttling, and may
render lots of discarded frames.

It is an alternative solution to solving
`VUID-vkAcquireNextImageKHR-semaphore-01779` by also using a unique
`Semaphore` that we know is no longer waited on via previously waiting
on the relevant `Fence`; although similar in nature to the previous
solution.